+++ CHANGELOG.md
@@ -19,9 +19,7 @@
       tracking and error reporting and are available for review in the Admin Console.
     * New `XH.genUUID()` method to generate UUIDs for use as Correlation IDs or elsewhere.
 * `GridModel.contextMenu` will now accept `false` as a value to omit context menus.
-*  New global event handlers on `FetchService`. See `FetchService.addExceptionHandler()` and
-   `FetchService.addSuccessHandler()`.
-
+*  New global exception handlers on `FetchService`. See `FetchService.addExceptionHandler()`.
 
 ### 🐞 Bug Fixes